Renamed update related classes
